requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
torch>=1.9.0
pytorch-lightning>=1.4.0
pyyaml>=6.0
plotly>=5.3.0
yfinance>=0.1.70
tqdm>=4.62.0
requests>=2.25.0
python-dotenv>=1.0.1
